Added more compile flags and fixed all compile warnings

+++ src/gui/Font.cc
@@ -36,7 +36,7 @@ Font::Font(std::string const& font_name, int font_size, vec3 const& color)
     std::getline(font_file, curline);
     line_values = get__values_from_string(curline);
 
-    vec2 font_image_size {line_values["scaleW"], line_values["scaleH"]};
+    vec2 font_image_size {static_cast<float>(line_values["scaleW"]), static_cast<float>(line_values["scaleH"])};
     float base_line {line_values["base"] / font_image_size.y * font_size / default_font_size};
 
     std::getline(font_file, curline);
